add img, rename file, update homepage,

+++ SFCS/src/dao/impl/CartItemDaoImpl.java
@@ -61,7 +61,7 @@ public class CartItemDaoImpl extends JDBCConnection implements CartItemDao {
 			ps.setInt(1, cartItem.getCart().getId());
 			ps.setInt(2, cartItem.getProduct().getId());
 			ps.setInt(3, cartItem.getQuantity());
-			ps.setInt(4, cartItem.getUnitPrice());
+			ps.setLong(4, cartItem.getUnitPrice());
 			ps.setInt(5, cartItem.getId());		
 			
 			ps.executeUpdate();
